Moved TypePropagationTrace to separate file, made its methods elidable.
